move input logic MLSS to other file, and extract table format to yet another file
