+++ client/src/pages/Ramptest.jsx
@@ -1,6 +1,22 @@
+import { useState } from "react";
+import App from "../App"
 
 function Ramptest() {
-    return <h1>Page to analyse a ramp test based on speed/power versus heart rate/lactate.</h1>
+    const[x, setx] = useState();
+    return (
+    <>
+    <h1>Page to analyse a ramp test based on speed/power versus heart rate/lactate.</h1>
+    <input
+        type="number"
+        placeholder="speed"
+        id="one"
+        value={x}
+        onChange={e => setx(e.target.value)}
+    />
+    <input /><input /><input /><input /><br/>
+   
+    </>
+    )
 }
 
 export default Ramptest;
